Move iAP into Ice
