+++ cpp/include/Ice/iAP/ConnectionInfo.h
