+++ cpp/include/Ice/iAP/EndpointInfo.h
