--- cpp/include/IceIAP/IceIAP.h
@@ -1,11 +1,0 @@
-//
-// Copyright (c) ZeroC, Inc. All rights reserved.
-//
-
-#ifndef ICE_IAP_ICE_IAP_H
-#define ICE_IAP_ICE_IAP_H
-
-#include "ConnectionInfo.h"
-#include "EndpointInfo.h"
-
-#endif
