+++ cpp/src/Ice/iAP/EndpointI.mm
@@ -6,12 +6,12 @@
 
 #if TARGET_OS_IPHONE != 0
 
-#    include "../Ice/DefaultsAndOverrides.h"
-#    include "../Ice/EndpointFactoryManager.h"
-#    include "../Ice/HashUtil.h"
-#    include "../Ice/Network.h"
-#    include "../Ice/ProtocolInstance.h"
-#    include "../Ice/ProtocolPluginFacade.h"
+#    include "../DefaultsAndOverrides.h"
+#    include "../EndpointFactoryManager.h"
+#    include "../HashUtil.h"
+#    include "../Network.h"
+#    include "../ProtocolInstance.h"
+#    include "../ProtocolPluginFacade.h"
 #    include "Connector.h"
 #    include "EndpointI.h"
 #    include "Ice/Initialize.h"
@@ -20,8 +20,8 @@
 #    include "Ice/OutputStream.h"
 #    include "Ice/Properties.h"
 #    include "Ice/RegisterPlugins.h"
-#    include "IceIAP/ConnectionInfo.h"
-#    include "IceIAP/EndpointInfo.h"
+#    include "Ice/iAP/ConnectionInfo.h"
+#    include "Ice/iAP/EndpointInfo.h"
 
 #    include <CoreFoundation/CoreFoundation.h>
